Add files via upload
Initially, we possess a dataset in the OWL file format. 
Subsequently, we transform this .OWL file into a format suitable for input into R and Python. 
Following this, we apply various algorithms to our dataset, which involves preprocessing. 
These algorithms are implemented in  R and Python languages and encompass LDA, LSA, BERTopic, and Top2vec.